Work Log #7
I had a ton of trouble getting the graphics to work, so I scrapped the
graphics portion and is basing the game off a basic array. After I make
the algorithm and have the game work, I will try to implement the
graphics portion. Today I spent the class getting making the shuffle
method.